feat: добавлена работа с редис в сервис оплаты. Контроллер в сервисе магазина разбит на несколько
