+++ payment-service/src/main/java/ru/yandex/practicum/eshop/payment/service/exceptions/ItemNotFoundException.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.eshop.payment.service.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+/**
+ * Исключение в случае, если пришел некорректный тип сортировки.
+ */
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class ItemNotFoundException extends RuntimeException {
+  public ItemNotFoundException(String message) {
+    super(message);
+  }
+}
